+++ internal/common/common.go
@@ -1,7 +1,5 @@
 package common
 
-import "slices"
-
 // MapToSlice transforms a map into a slice based on specific iteratee
 // Play: https://go.dev/play/p/ZuiCZpDt6LD
 func MapToSlice[K comparable, V any, R any](in map[K]V, iteratee func(key K, value V) R) []R {
@@ -12,11 +10,6 @@ func MapToSlice[K comparable, V any, R any](in map[K]V, iteratee func(key K, val
 	}
 
 	return result
-}
-
-// Contains returns true if an element is present in a collection.
-func Contains[T comparable](collection []T, element T) bool {
-	return slices.Contains(collection, element)
 }
 
 // ToAnySlice returns a slice with all elements mapped to `any` type